removed Head / added metadata
