+++ app/projects/(single projects)/glasstic/page.js
@@ -5,7 +5,9 @@ import SocialMedia from "./ui/SocialMedia";
 import ProjectHero from "@/components/ProjectHero";
 import Result from "@/components/Result";
 import Outro from "@/components/Outro";
-import { getNextProject } from "@/utils/data/getData";
+import { getNextProject, getProjectMeta } from "@/utils/data/getData";
+
+export const metadata = getProjectMeta("glasstic");
 
 export default async function Page() {
   const project = await getNextProject();
