+++ app/projects/(single projects)/imunoshop/page.js
@@ -8,7 +8,9 @@ import Banner from "./ui/Banner";
 import ProjectHero from "@/components/ProjectHero";
 import Result from "@/components/Result";
 import Outro from "@/components/Outro";
-import { getNextProject } from "@/utils/data/getData";
+import { getNextProject, getProjectMeta } from "@/utils/data/getData";
+
+export const metadata = getProjectMeta("imunoshop");
 
 export default async function Page() {
   const project = await getNextProject();
