+++ app/projects/(single projects)/joker/page.js
@@ -8,7 +8,9 @@ import Roll from "./ui/Roll";
 import ProjectHero from "@/components/ProjectHero";
 import Result from "@/components/Result";
 import Outro from "@/components/Outro";
-import { getNextProject } from "@/utils/data/getData";
+import { getNextProject, getProjectMeta } from "@/utils/data/getData";
+
+export const metadata = getProjectMeta("joker");
 
 export default async function Page() {
   const project = await getNextProject();
