+++ app/projects/(single projects)/klei/page.js
@@ -6,7 +6,9 @@ import Hands from "./ui/Hands";
 import ProjectHero from "@/components/ProjectHero";
 import Result from "@/components/Result";
 import Outro from "@/components/Outro";
-import { getNextProject } from "@/utils/data/getData";
+import { getNextProject, getProjectMeta } from "@/utils/data/getData";
+
+export const metadata = getProjectMeta("klei");
 
 export default async function Page() {
   const project = await getNextProject();
