+++ app/projects/(single projects)/northprim/page.js
@@ -1,4 +1,3 @@
-import Head from "next/head";
 import VideoSection from "./ui/VideoSection";
 import Brand from "./ui/Brand";
 import Sprint from "./ui/Sprint";
@@ -16,7 +15,9 @@ import Coffee from "./ui/Coffee";
 import ProjectHero from "@/components/ProjectHero";
 import Result from "@/components/Result";
 import Outro from "@/components/Outro";
-import { getNextProject } from "@/utils/data/getData";
+import { getNextProject, getProjectMeta } from "@/utils/data/getData";
+
+export const metadata = getProjectMeta("northprim");
 
 export default async function Page() {
   const project = await getNextProject();
