+++ app/projects/(single projects)/ownnew/page.js
@@ -6,7 +6,9 @@ import WebDevelopment from "./ui/WebDevelopment";
 import ProjectHero from "@/components/ProjectHero";
 import Result from "@/components/Result";
 import Outro from "@/components/Outro";
-import { getNextProject } from "@/utils/data/getData";
+import { getNextProject, getProjectMeta } from "@/utils/data/getData";
+
+export const metadata = getProjectMeta("ownnew");
 
 export default async function Page() {
   const project = await getNextProject();
