+++ app/projects/(single projects)/zeder/page.js
@@ -5,7 +5,9 @@ import Banner from "./ui/Banner";
 import ProjectHero from "@/components/ProjectHero";
 import Result from "@/components/Result";
 import Outro from "@/components/Outro";
-import { getNextProject } from "@/utils/data/getData";
+import { getNextProject, getProjectMeta } from "@/utils/data/getData";
+
+export const metadata = getProjectMeta("zeder");
 
 export default async function Page() {
   const project = await getNextProject();
